Moving CLI class into the main config file
